feat: :sparkles: enhance PictureCanvas with custom hook and improved image handling
Refactored the PictureCanvas component to utilize a new custom hook, `usePictureCanvas`, for better state management and image handling. Added error handling and loading states for image uploads. Updated the PictureViewer component with enhanced documentation and improved accessibility. Adjusted styles for a more cohesive user experience and enabled TypeScript checks in the project configuration.
