+++ src/app/editor/components/PictureCanvas/style.scss
@@ -15,7 +15,6 @@
   &-image {
     width: 100%;
     height: 100%;
-    border-radius: v.$radius;
     object-fit: cover;
   }
