+++ src/app/editor/workers/upload.worker.ts
