+++ tsconfig.json
@@ -4,6 +4,7 @@
     "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
+    "checkJs": true,
     "strict": true,
     "noEmit": true,
     "esModuleInterop": true,
